:bug: Fix home playlists section, loading on playlists page, validation
